// packages/octez.connect-blockchain-tezos-sapling/src/types/messages/permission-response.ts
import { AppMetadata, NetworkType, PermissionResponseV3 } from '@tezos-x/octez.connect-types'
import { TezosSaplingPermissionScope } from '../permission-scope'

export interface TezosSaplingPermissionResponse extends PermissionResponseV3<'tezos-sapling'> {
  blockchainData: {
    appMetadata: AppMetadata
    scopes: TezosSaplingPermissionScope[] // enum
    accounts: {
      accountId: string
      address: string
      viewingKey?: string // If the "viewing key" scope is not set, this value has to be removed by the SDK
      network: {
        contract: string // sapling contract
        type: NetworkType
        name?: string
        rpcUrl?: string
      }
    }[]
  }
}